Final refactor of stylesheets and components
