+++ src/styles/atoms/PeopleComponent.module.css
